requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
openai
langchain-openai
langchain-core
chromadb
fitz  # PyMuPDF
asyncio
re
